LayoutTests/fast/forms/datetimelocal/datetimelocal-editable-components/accessibility-opens-picker-expected.txt

Tests that an accessibility press can activate the date picker.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS showingPicker is false
PASS successfullyParsed is true

TEST COMPLETE
